requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
attrs==19.3.0
boto3==1.11.9
botocore==1.14.9          # via boto3, s3transfer
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.0                # via flask
docutils==0.15.2          # via botocore
flask==1.1.1
humanize==0.5.1
idna==2.8                 # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.11.0            # via flask
jmespath==0.9.4           # via boto3, botocore
markupsafe==1.1.1         # via jinja2
oauthlib==3.1.0           # via requests-oauthlib
python-dateutil==2.8.1    # via botocore
requests-oauthlib==1.3.0  # via wellcome-storage-service
requests==2.22.0          # via requests-oauthlib
s3transfer==0.3.2         # via boto3
six==1.14.0               # via python-dateutil
tqdm==4.42.0
urllib3==1.25.8           # via botocore, requests
wellcome-storage-service==1.5.0
werkzeug==0.16.1          # via flask
zipstreamer==0.1.3
